Add type guard utilities (#91)
Adds some type guard utilities in a new file, `utils.ts`, and accompanying unit tests.

In addition to fixing #89, adds a function for validating JSON-RPC ID values. As part of implementing that, this also fixes a long-standing bug where we returned `undefined` instead of `null` for the ID when returning an error for a request without an `id` value.